+++ 会議室予約システムAPI/sql_app/main.py
@@ -1,0 +1,50 @@
+from typing import List
+from fastapi import FastAPI ,Depends
+from sqlalchemy.orm import Session
+
+from . import crud, models, schemas
+from .database import SessionLocal, engine
+
+models.Base.metadata.create_all(bind=engine)
+
+app = FastAPI()
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
+
+# @app.get("/")
+# async def index():
+#     return {"message" : "Success"}
+
+# Read
+@app.get("/users", response_model=List[schemas.User])
+async def read_users(skip: int = 0, limit: int = 100, db: Session = Depends(get_db)):
+    users = crud.get_users(db, skip=skip, limit=limit)
+    return users
+
+@app.get("/room", response_model=List[schemas.Room])
+async def read_room(skip: int = 0, limit: int = 100, db: Session = Depends(get_db)):
+    rooms = crud.get_rooms(db, skip=skip, limit=limit)
+    return rooms
+
+@app.get("/booking", response_model=List[schemas.Booking])
+async def read_booking(skip: int = 0, limit: int = 100, db: Session = Depends(get_db)):
+    bookings = crud.get_bookings(db, skip=skip, limit=limit)
+    return bookings
+
+# Create
+@app.post("/users")
+async def users(users: User):
+    return {"users" : users}
+
+@app.post("/rooms")
+async def rooms(rooms: Room):
+    return {"rooms" : rooms}
+
+@app.post("/bookings")
+async def bookings(bookings: Booking):
+    return {"bookings" : bookings}
